@@ -2,7 +2,7 @@
   import Highlight from "svelte-highlight";
   import javascript from "svelte-highlight/languages/javascript";
 
-  export let data;
+  let { data } = $props();
 </script>
 
 <Highlight language={javascript} code=
